requirements.txt
GitPython==3.1.31
gitdb
typing-extensions
pydriller
dulwich
PyYAML